amo-backend/requirements.txt
django-auto-prefetching==0.1.8
Django==3.2.4
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.3.0
drf-yasg==1.20.0
gunicorn==20.1.0
model_bakery==1.11.0
numpy==1.22.2
pandas==1.4.2
Pillow==9.1.1
PyJWT==1.7.1
python-dateutil==2.8.1
requests==2.26.0
sqlparse==0.3.1
urllib3==1.26.9